+++ cubano-webdriver-manager/src/main/java/org/concordion/cubano/driver/web/provider/EdgeBrowserProvider.java
@@ -6,36 +6,37 @@ import org.concordion.cubano.driver.web.config.WebDriverConfig;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.edge.EdgeOptions;
+
 import io.github.bonigarcia.wdm.EdgeDriverManager;
 
 public class EdgeBrowserProvider extends LocalBrowserProvider {
-	public static final String BROWSER_NAME = "edge";
-    
+    public static final String BROWSER_NAME = "edge";
+
     /**
      * @return Starts Edge driver manager and creates a new WebDriver instance.
      */
-	@Override
-	public WebDriver createDriver() {
-	    	setupBrowserManager(EdgeDriverManager.getInstance());
-	
-	    	EdgeOptions options = new EdgeOptions();
-        
+    @Override
+    public WebDriver createDriver() {
+        setupBrowserManager(EdgeDriverManager.getInstance());
+
+        EdgeOptions options = new EdgeOptions();
+
         addProxyCapabilities(options);
         addCapabilities(options);
-        
+
         WebDriver driver = new EdgeDriver(options);
-        
+
         setBrowserSize(driver);
-        
+
         return driver;
     }
 
-	private void addCapabilities(EdgeOptions options) {
-	    	Map<String, String> settings = WebDriverConfig.getInstance().getPropertiesStartingWith("edge.capability.", true);
-	    	
-	    	for (String key : settings.keySet()) {
-	   		options.setCapability(key, settings.get(key));
-		}
-	}
-    
+    private void addCapabilities(EdgeOptions options) {
+        Map<String, String> settings = WebDriverConfig.getInstance().getPropertiesStartingWith("edge.capability.", true);
+
+        for (String key : settings.keySet()) {
+            options.setCapability(key, settings.get(key));
+        }
+    }
+
 }
